+++ projects/demo/src/app/app.config.ts
@@ -1,9 +1,16 @@
 import { ApplicationConfig } from '@angular/core';
 import { provideRouter } from '@angular/router';
 
-import { routes } from './app.routes';
 import { provideClientHydration } from '@angular/platform-browser';
+import { provideNgxCountAnimations } from 'ngx-count-animation';
+import { routes } from './app.routes';
 
 export const appConfig: ApplicationConfig = {
-  providers: [provideRouter(routes), provideClientHydration()]
+  providers: [
+    provideRouter(routes),
+    provideClientHydration(),
+    provideNgxCountAnimations({
+      /// Add your global configs
+    }),
+  ]
 };