+++ projects/ngx-count-animation/src/lib/ngx-count-animation.module.ts
@@ -1,12 +1,34 @@
-import { NgModule } from '@angular/core';
+import { InjectionToken, ModuleWithProviders, NgModule, Provider } from '@angular/core';
 
 import { NgxCountAnimationDirective } from './ngx-count-animation.directive';
-import { NgxCountService } from './ngx-count-animation.service';
+import { NgxCountAnimationConfigs } from './utils/coercion/ngx-count-animation-configs';
+
+export const NGX_COUNT_ANIMATION_CONFIGS =
+    new InjectionToken<NgxCountAnimationConfigs>('NgxCountAnimationConfigs');
+
 
 @NgModule({
-    imports: [],
+    imports: [NgxCountAnimationDirective],
     exports: [NgxCountAnimationDirective],
-    declarations: [NgxCountAnimationDirective],
-    providers: [NgxCountService],
 })
-export class NgxCountAnimationModule { }
+export class NgxCountAnimationModule {
+    static forRoot(
+        configs: NgxCountAnimationConfigs
+    ): ModuleWithProviders<NgxCountAnimationModule> {
+        return {
+            ngModule: NgxCountAnimationModule,
+            providers: [provideNgxCountAnimations(configs)],
+        };
+    }
+}
+
+export function provideNgxCountAnimations(
+    configs?: NgxCountAnimationConfigs
+): Provider {
+    return [
+        {
+            provide: NGX_COUNT_ANIMATION_CONFIGS,
+            useValue: configs,
+        },
+    ];
+}